attempt at webpack
