+++ .eslintrc
@@ -1,0 +1,188 @@
+{
+  "env": {
+    "browser": true,
+    "node": true,
+    "jasmine": true
+  },
+  "globals": {
+      "$": true
+  },
+  "rules": {
+    "comma-dangle" : [2, "never"],
+    "no-cond-assign": [2],
+    "no-console": [2],
+    "no-constant-condition": [2],
+    "no-control-regex": [2],
+    "no-debugger": [2],
+    "no-dupe-args": [2],
+    "no-dupe-keys": [2],
+    "no-duplicate-case": [2],
+    "no-empty-character-class": [2],
+    "no-empty": [2],
+    "no-ex-assign": [2],
+    "no-extra-boolean-cast": [0],
+    "no-extra-parens": [2, "functions"],
+    "no-extra-semi": [2],
+    "no-func-assign": [2],
+    "no-inner-declarations": [2, "both"],
+    "no-invalid-regexp": [2],
+    "no-irregular-whitespace": [2],
+    "no-negated-in-lhs": [2],
+    "no-obj-calls": [2],
+    "no-regex-spaces": [2],
+    "no-sparse-arrays": [2],
+    "no-unreachable": [2],
+    "use-isnan": [2],
+    "valid-jsdoc": [0],
+    "valid-typeof": [2],
+    "no-unexpected-multiline": [2],
+
+    "accessor-pairs": [0],
+    "block-scoped-var": [2],
+    "complexity": [2, 5],
+    "consistent-return": [2],
+    "curly": [2],
+    "default-case": [0],
+    "dot-notation": [2],
+    "dot-location": [2, "property"],
+    "eqeqeq": [2],
+    "guard-for-in": [2],
+    "no-alert": [2],
+    "no-caller": [2],
+    "no-div-regex": [2],
+    "no-else-return": [2],
+    "no-eq-null": [2],
+    "no-eval": [2],
+    "no-extend-native": [2],
+    "no-extra-bind": [2],
+    "no-fallthrough": [2],
+    "no-floating-decimal": [2],
+    "no-implicit-coercion": [0],
+    "no-implied-eval": [2],
+    "no-invalid-this": [0],
+    "no-iterator": [2],
+    "no-labels": [2, {"allowLoop": false, "allowSwitch": false}],
+    "no-lone-blocks": [2],
+    "no-loop-func": [2],
+    "no-multi-spaces": [2],
+    "no-multi-str": [2],
+    "no-native-reassign": [2],
+    "no-new-func": [2],
+    "no-new-wrappers": [2],
+    "no-new": [2],
+    "no-octal-escape": [2],
+    "no-octal": [2],
+    "no-param-reassign": [2],
+    "no-process-env": [0],
+    "no-proto": [2],
+    "no-redeclare": [2],
+    "no-return-assign": [2],
+    "no-script-url": [2],
+    "no-self-compare": [2],
+    "no-sequences": [2],
+    "no-throw-literal": [2],
+    "no-unused-expressions": [2],
+    "no-useless-call": [2],
+    "no-useless-concat": [2],
+    "no-void": [2],
+    "no-warning-comments": [2],
+    "no-with": [2],
+    "radix": [2],
+    "vars-on-top": [2],
+    "wrap-iife": [2],
+    "yoda": [0],
+
+    "strict": [0, "global"],
+
+    "init-declarations": [0],
+    "no-catch-shadow": [2],
+    "no-delete-var": [2],
+    "no-label-var": [2],
+    "no-shadow-restricted-names": [2],
+    "no-shadow": [2],
+    "no-undef-init": [2],
+    "no-undef": [2],
+    "no-undefined": [2],
+    "no-unused-vars": [2],
+    "no-use-before-define": [2],
+
+    "callback-return": [2],
+    "global-require": [2],
+    "handle-callback-err": [2],
+    "no-mixed-requires": [0],
+    "no-new-require": [2],
+    "no-path-concat": [2],
+    "no-process-exit": [2],
+    "no-restricted-modules": [0],
+    "no-sync": [0],
+
+    "array-bracket-spacing": [2, "never"],
+    "block-spacing": [2],
+    "brace-style": [2],
+    "camelcase": [2],
+    "comma-spacing": [2, {"before": false, "after": true}],
+    "comma-style": [2],
+    "computed-property-spacing": [2],
+    "consistent-this": [0, "self"],
+    "eol-last": [2],
+    "func-names": [0],
+    "func-style": [2,  "declaration"],
+    "id-length": [0],
+    "id-match": [0],
+    "indent": [2, 4, {
+      "VariableDeclarator": 1
+    }],
+    "jsx-quotes": [0],
+    "key-spacing": [0],
+    "lines-around-comment": [2],
+    "linebreak-style": [0],
+    "max-nested-callbacks": [2, 5],
+    "new-cap": [2],
+    "new-parens": [2],
+    "newline-after-var": [2, "always"],
+    "no-array-constructor": [2],
+    "no-continue": [0],
+    "no-inline-comments": [2],
+    "no-lonely-if": [2],
+    "no-mixed-spaces-and-tabs": [2],
+    "no-multiple-empty-lines": [2],
+    "no-nested-ternary": [2],
+    "no-new-object": [2],
+    "no-restricted-syntax": [2],
+    "no-spaced-func": [2],
+    "no-ternary": [0],
+    "no-trailing-spaces": [2],
+    "no-underscore-dangle": [2],
+    "no-unneeded-ternary": [2],
+    "object-curly-spacing": [2, "never"],
+    "operator-assignment": [0],
+    "operator-linebreak": [2, "after"],
+    "padded-blocks": [0],
+    "quote-props": [0],
+    "quotes": [2, "single"],
+    "require-jsdoc": [0],
+    "semi-spacing": [2],
+    "semi": [2],
+    "sort-vars": [0],
+    "keyword-spacing": [2, {"before": true, "after": true, "overrides": {
+      "if": {"after": true},
+      "for": {"after": true},
+      "while": {"after": true},
+      "throw": {"after": true}
+    }}],
+    "space-before-blocks": [2],
+    "space-before-function-paren": [2, {"anonymous": "always", "named": "never"}],
+    "space-in-parens": [2],
+    "space-infix-ops": [2],
+    "space-unary-ops": [2],
+    "spaced-comment": [2],
+    "wrap-regex": [2],
+
+    "max-depth": [2],
+    "max-len": [2, 120],
+    "max-params": [0, 3],
+    "max-statements": [0],
+    "no-bitwise": [2],
+    "no-plusplus": [2, { "allowForLoopAfterthoughts": true }]
+  }
+}
