+++ .gitignore
@@ -1,0 +1,5 @@
+.idea*
+*.sass-cache*
+node_modules
+npm-debug.log
+dist
