+++ README.txt
@@ -1,2 +1,29 @@
-Infinite Scroll Project
+# HeavenScroll
+> Heaven scroll provides an InfinateScroll with amazing performance and experiance improvements
 
+## Pre-requisites
+Install the package dependencies by running the following command
+```bash
+npm install
+```
+
+## Development
+```
+npm run start
+```
+
+Then view this project at http://localhost:8080
+
+## Demo
+To create a packagable demo, use the following command.
+```bash
+npm run build
+```
+
+This will create a demo project in the dist folder
+
+## Publishing
+```
+npm run build
+npm publish
+```
