+++ demo/index.html
@@ -4,9 +4,7 @@
 		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width">
 		<title>Infinite Scroll Test</title>
-		<link rel="stylesheet" href="style.css">
-		<script src="localLib/jQuery.js"></script>
-		<script src="localLib/underscore.js"></script>
+		<link rel="stylesheet" href="./demo.css">
 	</head>
 	<body>
 		<div class="mainDiv">
@@ -30,6 +28,8 @@
 			<p>Deleted Pages: <span class="deletedPageValue"></span></p>
 		</div>
 
-		<script src="scripts.js"></script>
+		<script src="dependencies.js"></script>
+		<script src="HeavenScroll.js"></script>
+		<script src="demo.js"></script>
 	</body>
 </html>
