+++ package.json
@@ -1,0 +1,56 @@
+{
+  "name": "jquery-heaven-scroll",
+  "version": "1.0.0",
+  "description": " Heaven scroll provides an InfinateScroll with amazing performance and experiance improvements",
+  "main": "dist/module.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack-dev-server --progress --colors --content-base build/",
+    "build": "webpack --progress --colors",
+    "eslint": "eslint"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Comum/InfiniteScroll.git"
+  },
+  "keywords": [
+    "jquery",
+    "webpack",
+    "scss"
+  ],
+  "author": "Muhammad Dadu",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Comum/InfiniteScroll/issues"
+  },
+  "homepage": "https://github.com/Comum/InfiniteScroll#readme",
+  "devDependencies": {
+    "autoprefixer": "6.7.5",
+    "babel-core": "6.23.1",
+    "babel-loader": "6.3.2",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "babel-preset-es2015": "6.22.0",
+    "bootstrap": "3.3.7",
+    "bootstrap-sass": "3.3.7",
+    "copy-webpack-plugin": "4.0.1",
+    "css-loader": "0.26.1",
+    "eslint": "3.19.0",
+    "eslint-loader": "1.7.1",
+    "extract-text-webpack-plugin": "2.0.0-rc.3",
+    "file-loader": "0.10.0",
+    "image-webpack-loader": "3.2.0",
+    "imports-loader": "0.7.0",
+    "jquery": "3.1.1",
+    "lodash": "4.17.4",
+    "node-sass": "4.5.0",
+    "postcss-loader": "1.3.2",
+    "resolve-url-loader": "2.0.0",
+    "sass-loader": "6.0.2",
+    "sass-mq": "3.3.2",
+    "script-loader": "0.7.0",
+    "style-loader": "0.13.1",
+    "webpack": "2.2.1",
+    "webpack-dev-server": "^2.7.1"
+  }
+}
